% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spax_nmf.R
\name{basis}
\alias{basis}
\title{NMF-style Basis Matrix}
\usage{
basis(object, ...)
}
\arguments{
\item{object}{Result from [spax_nmf()].}

\item{...}{Unused.}
}
\value{
A numeric matrix with wavelengths in rows and components in columns.
}
\description{
Returns the component spectra arranged as wavelength-by-component, similar to
a basis matrix in NMF software interfaces.
}
\examples{
fake <- list(spectra = matrix(runif(15), 3, 5))
class(fake) <- "spax_nmf"
basis(fake)
}
