% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spax_nmf.R
\name{coef.spax_nmf}
\alias{coef.spax_nmf}
\title{Extract Component-Weight Coefficients}
\usage{
\method{coef}{spax_nmf}(object, ...)
}
\arguments{
\item{object}{Result from [spax_nmf()].}

\item{...}{Unused.}
}
\value{
A numeric matrix with components in rows and spaxels in columns.
}
\description{
Returns the component weights arranged as component-by-spaxel, similar to the
coefficient matrix used by NMF interfaces.
}
\examples{
fake <- list(spatial = matrix(runif(12), 4, 3))
class(fake) <- "spax_nmf"
coef(fake)
}
