% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spax_nmf.R
\name{component_spectrum}
\alias{component_spectrum}
\title{Extract a Component Spectrum}
\usage{
component_spectrum(fit, component = 1)
}
\arguments{
\item{fit}{Result from [spax_nmf()].}

\item{component}{Component index to extract.}
}
\value{
A numeric vector.
}
\description{
Returns one recovered component spectrum from a fitted model.
}
\examples{
\dontrun{
cube <- simulate_ifu_cube(nx = 6, ny = 5, n_wave = 20)
fit <- spax_nmf(cube$matrix, k = 3, niter = 50)
spec1 <- component_spectrum(fit, 1)
}
}
