% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spax_nmf.R
\name{plot_reconstruction}
\alias{plot_reconstruction}
\title{Plot Data Versus Reconstruction}
\usage{
plot_reconstruction(fit, x, pixels = NULL, n = 6, wavelength = NULL, ...)
}
\arguments{
\item{fit}{Result from [spax_nmf()].}

\item{x}{Numeric matrix used as input to the fit.}

\item{pixels}{Optional vector of spaxel indices to plot.}

\item{n}{Number of random spaxels to plot when \code{pixels} is omitted.}

\item{wavelength}{Optional wavelength vector.}

\item{...}{Additional graphical arguments passed to \code{plot()}.}
}
\value{
Invisibly returns \code{fit}.
}
\description{
Plots observed and reconstructed spectra for selected spaxels.
}
\examples{
\dontrun{
cube <- simulate_ifu_cube(nx = 6, ny = 5, n_wave = 20)
fit <- spax_nmf(cube$matrix, k = 3, niter = 50)
plot_reconstruction(fit, cube$matrix, n = 4, wavelength = cube$wavelength)
}
}
