% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaxNMF-package.R
\docType{package}
\name{SpaxNMF-package}
\alias{SpaxNMF}
\alias{SpaxNMF-package}
\title{SpaxNMF: Regularized Non-negative Matrix Factorization for IFU Cubes}
\description{
Tools for reshaping IFU cubes, fitting smooth non-negative factorization
models, and building reproducible examples for package documentation and
comparison experiments.
}
\details{
The package is designed around a simple workflow:
\enumerate{
  \item simulate or ingest a spectral cube;
  \item reshape it with [cube_to_matrix()];
  \item fit a low-rank unmixing model with [spax_nmf()];
  \item inspect component spectra and component-weight maps.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RafaelSdeSouza/SpaxNMF}
  \item Report bugs at \url{https://github.com/RafaelSdeSouza/SpaxNMF/issues}
}

}
\author{
\strong{Maintainer}: Rafael S. de Souza \email{r.s.desouza@herts.ac.uk} (\href{https://orcid.org/0000-0001-7207-4584}{ORCID})

}
\keyword{internal}
