% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spax_nmf.R
\name{summary.spax_nmf}
\alias{summary.spax_nmf}
\title{Summarize a SpaxNMF Fit}
\usage{
\method{summary}{spax_nmf}(object, ...)
}
\arguments{
\item{object}{Result from [spax_nmf()].}

\item{...}{Unused.}
}
\value{
A list with class \code{"summary.spax_nmf"}.
}
\description{
Summarize a SpaxNMF Fit
}
\examples{
fake <- list(
  spatial = matrix(runif(12), 4, 3),
  spectra = matrix(runif(15), 3, 5),
  loss = c(3, 2, 1)
)
class(fake) <- "spax_nmf"
summary(fake)
}
