% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spax_nmf.R
\name{component_map}
\alias{component_map}
\title{Extract a Component Map}
\usage{
component_map(fit, nx, ny, component = 1)
}
\arguments{
\item{fit}{Result from [spax_nmf()].}

\item{nx}{Number of x-axis pixels.}

\item{ny}{Number of y-axis pixels.}

\item{component}{Component index to extract.}
}
\value{
A numeric matrix of size \code{nx x ny}.
}
\description{
Reshapes one component-weight column from a fitted model into an image plane.
}
\examples{
cube <- simulate_ifu_cube(nx = 6, ny = 5, n_wave = 20)
fit <- spax_nmf(cube$matrix, k = 3, niter = 50)
img <- component_map(fit, cube$nx, cube$ny, 2)
dim(img)
}
